requirements.txt
flask == 3.0.3
bootstrap-flask == 2.4.0
flask-sslify == 0.1.5
pandas == 2.0.3
openpyxl == 3.1.2
google-cloud-storage == 2.16.0
tqdm == 4.66.2